Woop Woop, I created a class Deck which has enough functions to deal a hand of texas hold em to one person (at the moment)
Now to start writing the comparison functions for finding probabilities. I plan to use regular expressions to divide the cards into rank and suit for comparisons. Probably would have been better to use the card class, but this is written and I can refactor it if need be later.

Just happy to have written this while we are out of cell service and I was forced to think through each problem instead of relying on stack overflow.