Poker Odds Game
This project is my foray into development of a statistics based game. Being a noob and needing to practice hacking on something aside from a tutorial, this is my first self directed project. Most of this code was written without a data connection, so while there are many more elegant solutions out there, this is what I came up with on the fly.